fix(infra): move temporal/near_dup_calibration/write_class core/ -> shared/ (issue #406)
infrastructure/ importing core/ is a hard layer-rule violation
(module-inventory.md dependency table, CLAUDE.md § Code Style). Four
infrastructure/ files needed temporal_normalize.normalize_date_to_iso,
near_dup_calibration.SCAN_FLOOR/CandidatePair, or
write_class.classify_write_class directly (pg_store_write.py,
pg_store_near_dup.py, pg_store_memory_reheat.py, sqlite_store.py) and
could not legally import them from core/.

Root cause: temporal.py, temporal_normalize.py, temporal_timezones.py,
near_dup_calibration.py, and write_class.py are pure business-rule-free
utilities — stdlib-only imports (math/re/datetime,
collections.abc/typing, NamedTuple), zero I/O, zero dependency on any
other core/ module's business logic (temporal_normalize.py depends
only on its sibling temporal.py/temporal_timezones.py, which moved
with it). They satisfy shared/'s own stated criterion ("pure utility
functions... no dependencies on other project layers") and were
simply filed under the wrong layer.

Fix: relocated all five to shared/, updated every import site across
core/, infrastructure/, handlers/, tests_py/, scripts/, and
benchmarks/ (22 files) — both `from mcp_server.core.X import Y` and
`from mcp_server.core import X` forms. docs/module-inventory.md
updated: entries moved from the core/ section to shared/, dependency
counts re-measured and sourced.

tests_py/invariants/test_I2_canonical_writer.py: homeostatic_apply.py's
line-pinned heat_base writer shifted 233->234 (one new `from
mcp_server.shared import write_class` import line above the site).
Same writer, not new.

Verified: full suite green (see subsequent commits' final run), ruff
check + format clean, pyright zero-diagnostic on mcp_server/. No
functional change — every relocated symbol keeps its name and
signature; only the import path changes.
