+++ mcp_server/handlers/recall_helpers.py
@@ -18,7 +18,7 @@ from mcp_server.core.retrieval_signals import (
     compute_hopfield_hdc,
 )
 from mcp_server.core.scoring import compute_bm25_scores, compute_ngram_score
-from mcp_server.core.temporal import compute_recency_boost
+from mcp_server.shared.temporal import compute_recency_boost
 from mcp_server.infrastructure.embedding_engine import EmbeddingEngine
 from mcp_server.infrastructure.memory_store import MemoryStore
 from mcp_server.observability import silent_failure
