+++ mcp_server/shared/write_class.py
